requirements.txt
flask
pypdf
sentence-transformers
chromadb
openai
python-dotenv